+++ inst/unitTests/test_mvBIC.R
@@ -1,6 +1,10 @@
 # Gabriel Hoffman
 # April 12, 2020
 
+library(BiocParallel)
+register(SerialParam())
+
+
 # The values of BIC and evalCriterion are different
 # but the *difference* in values for two models are the same
 test_mvBIC = function(){
@@ -67,7 +71,7 @@ test_multiple_mbBIC_random = function(){
 	fit2 = lme4::lmer(Sepal.Length ~ (1|Species), data=iris, REML=FALSE )
 
 	scoreC = mvIC( list(fit1, fit2) )
-	scoreD = mvIC_fit( Y, ~ (1|Species), data=iris, fastApprox=FALSE)
+	scoreD = mvIC_fit( Y, ~ (1|Species), data=iris, pca=FALSE)
 
 	# scoreC@params$dataTerm
 	# scoreD@params$dataTerm
@@ -93,7 +97,7 @@ test_new_settings = function(){
 	c = mvIC( fit)	
 
 	fit = lm( Y ~ Species, data=iris)
-	d = mvIC( fit, fastApprox=TRUE)	
+	d = mvIC( fit, pca=FALSE)	
 
 	checkEqualsNumeric( a, b) & 
 	checkEqualsNumeric( a, c) & 
@@ -111,7 +115,7 @@ test_new_settings_mvIC_fit = function(){
 
 	b = mvIC_fit( t(Y), ~ Species, data=iris)
 
-	c = mvIC_fit( t(Y), ~ Species, data=iris, fastApprox=TRUE)
+	c = mvIC_fit( t(Y), ~ Species, data=iris, pca=FALSE)
 	   
 	checkEqualsNumeric( a, b) & 
 	checkEqualsNumeric( a, c)
@@ -120,15 +124,15 @@ test_new_settings_mvIC_fit = function(){
 
 test_new_settings_mvIC_fit_random = function(){ 
 
-	Y = with(iris, cbind(Sepal.Width, Sepal.Length))
-	rownames(Y) = rownames(iris)
+	Y = with(datasets::iris, cbind(Sepal.Width, Sepal.Length))
+	rownames(Y) = rownames(datasets::iris)
 
 	# test with fixed effects
 	variables = "Species"
-	bm = mvForwardStepwise( t(Y), ~ 1, data=iris, variables=variables, verbose=FALSE)
+	bm = mvForwardStepwise( t(Y), ~ 1, data=datasets::iris, variables=variables, verbose=FALSE)
 
-	a = mvIC_fit( t(Y), ~ 1, data=iris)
-	b = mvIC_fit( t(Y), ~ Species, data=iris)
+	a = mvIC_fit( t(Y), ~ 1, data=datasets::iris)
+	b = mvIC_fit( t(Y), ~ Species, data=datasets::iris)
 
 	a_score = with(a@params, dataTerm + penalty)
 	b_score = with(b@params, dataTerm + penalty)
@@ -137,10 +141,10 @@ test_new_settings_mvIC_fit_random = function(){
 
 	# test with random effects
 	variables = "(1|Species)"
-	bm = mvForwardStepwise( t(Y), ~ 1, data=iris, variables=variables, verbose=FALSE)
+	bm = mvForwardStepwise( t(Y), ~ 1, data=datasets::iris, variables=variables, verbose=FALSE)
 
-	a = mvIC_fit( t(Y), ~ 1, data=iris)
-	b = mvIC_fit( t(Y), ~ (1|Species), data=iris)
+	a = mvIC_fit( t(Y), ~ 1, data=datasets::iris)
+	b = mvIC_fit( t(Y), ~ (1|Species), data=datasets::iris)
 
 	a_score = with(a@params, dataTerm + penalty)
 	b_score = with(b@params, dataTerm + penalty)
@@ -150,11 +154,11 @@ test_new_settings_mvIC_fit_random = function(){
 	# Check formulas learned from exact and approximate methods
 	# --------
 	variables = c("Petal.Length", "Petal.Width", "(1|Species)")
+	
 	# exact
-	bm2 = mvForwardStepwise( t(Y), ~ 1, data=iris, variables=variables, verbose=FALSE)
+	bm2 = mvIC::mvForwardStepwise( t(Y), ~ 1, data=datasets::iris, variables=variables, verbose=FALSE)
 
-	# fastApprox is approximate for random effects
-	bm3 = mvForwardStepwise( t(Y), ~ 1, data=iris, variables=variables, fastApprox=TRUE, verbose=FALSE)
+	bm3 = mvForwardStepwise( t(Y), ~ 1, data=datasets::iris, variables=variables, pca=FALSE, verbose=FALSE)
 
 	checkEquals(bm2$formula, bm3$formula)
 }
@@ -162,6 +166,37 @@ test_new_settings_mvIC_fit_random = function(){
 
 
 
+test_check_accuray = function(){
+
+	set.seed(1)
+
+	# high dimensional data
+	n = 100
+	p = 2000
+	m = 15
+
+	X = matrix(rnorm(n*m), n,m)
+
+	beta = c(1,2,1)
+	Beta = matrix(beta, nrow=p, ncol=3)
+
+	Y = X[,1:3] %*% t(Beta) + matrix(rnorm(n*p), n,p)
+
+
+	Y = apply(Y, 2, function(x) x * rgamma(1,  1, 1e-3))
+
+	fit = mvForwardStepwise( t(Y), ~1, data.frame(X), colnames(data.frame(X)), verbose=FALSE)
+
+	# Y_scale = mvIC:::scale_features(Y)
+
+	# # very poor
+	# # fit = mvForwardStepwise( t(Y_scale), ~1, data.frame(X), colnames(data.frame(X)))
+
+	# # works but, less well
+	# # fitf = mvForwardStepwise( t(Y_scale), ~1, data.frame(X), colnames(data.frame(X)), fastApprox=TRUE)
+
+	checkEquals(sort(attr(terms(fit$formula), "term.labels")), c("X1", "X2", "X3") )
+}